+++ backend/src/Controller/User/GroupController.php
@@ -983,7 +983,7 @@ class GroupController extends BaseController
      *     ),
      *     @OA\Response(
      *         response="400",
-     *         description="This player is not a member of the required group(s)."
+     *         description="This player is not a member of one of the required groups."
      *     ),
      *     @OA\Response(
      *         response="403",